image2audio_lambda/requirements.txt
librosa
soundfile
matplotlib
pillow
sklearn
boto3